Extract and fix query unwrapped test
